requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pandas>=2.1.0
ccxt>=2.8.97
websockets>=11.0.3
python-dotenv>=1.0.0
